requirements.txt
beautifulsoup4==4.4.1
Django==1.7
django-appconf==1.0.1
django-generic-positions==0.1.7
django-geoposition==0.2.2
django-hvad==1.4.0
django-libs==1.66.12
django-ratings==0.3.7
django-rest-auth==0.7.0
django-review==1.9.2
django-user-media==1.1.6
djangorestframework==3.3.2
easy-thumbnails==2.3
factory-boy==2.6.0
fake-factory==0.5.3
geopy==1.11.0
oauthlib==1.0.3
Pillow==3.0.0
psycopg2==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.1
requests==2.9.1
requests-oauthlib==0.6.0
simplejson==3.8.1
six==1.10.0
South==1.0.2
wheel==0.24.0
